web: Move accounts and zones in blueprints
